requirements.txt
django
django-ckeditor
django-mptt